Control for the 6 motor added
I added new features to the reconbot package. The reconbot can be control it using thre different modes. full_mode, r_arm_mode and l_arm_mode. The first one allows to control the 2 arms at the same time and the r_arm_mode only takes control of the right arm and the l_arm_mode of the left one.